GitHub Actions CI workflow.
Ok, *THIS* one is lot more effort than other Workflows, as we have to
recompile mod_perl to ensure that it matches the version of Perl that we
are building/testnig with.

Worse yet, that _also_ requires that we build our own copy of Apache, as
the stock Ubuntu version has threads enabled (and all of our Perls are
non-threaded).

But wait, there's more!  Perls prior to 5.20 were not compiled as
shrplib out of the box, so we can't use those for testing here (unless
we want to recompile Perl as well).

Even more frustrating... Apache will refuse to start as `root`, so we
have to create our own `github` User to run the tests as, and have to
remember to `sudo --user github` when we want to do that.